services/alerts/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyzmq==25.1.1
PyYAML==6.0.1
redis==5.0.1
httpx==0.25.1
websockets==12.0
